requirements.txt
aiogram==3.13.0
telethon==1.36.0
groq==0.9.0
httpx==0.27.2
google-generativeai==0.8.3
aiohttp==3.10.0
aiosqlite==0.20.0
python-dotenv==1.0.1
elevenlabs==1.9.0
Pillow==10.4.0
